==> rev_fcst/s4_val_from_RF_and_cal_misp.ipynb <==
%%time 

import numpy as np
import pandas as pd
from dateutil.relativedelta import *
from pandas.tseries.offsets import *

pd.set_option('display.max_columns', None)
comp = pd.read_csv('~/misp_data/comp_1959-2019.csv', index_col=0)
crsp = pd.read_csv('~/misp_data/crsp_1959-2019.csv', index_col=0)
beta = pd.read_csv('~/misp_data/capm_beta_weekly_1-3y_1959-2019.csv', index_col=0)

comp['jdate'] = pd.to_datetime(comp['jdate'])
comp.permno = comp.permno.astype(int)

beta['date']=pd.to_datetime(beta['DATE'].astype(str))
beta['jdate']=beta['date']+MonthEnd(0)
beta['permno'] = beta.index
beta = beta.drop_duplicates(subset=['permno', 'jdate'])

FF = pd.read_csv('~/misp/lookup_tables/FF3.csv')
FF['date'] = FF['Unnamed: 0']
FF['date'] = pd.to_datetime(FF['date'].astype(str), format='%Y%m%d')
FF['jdate'] = FF['date']+MonthEnd(0)
FF = FF[['jdate', 'date', 'RF']]
FF['RF'] = (1+FF['RF']/100).pow(365)-1

beta = beta.merge(FF[['jdate', 'RF']],  on=['jdate'])
beta = beta.drop_duplicates(subset=['permno', 'jdate'])

def perform_valuation(df, val_method='dcf', **kwargs):
    if val_method == 'dcf':
        df['PV'] = dcf_valuation(df, **kwargs)
    else:
        raise ValueError(
            f'Invalid valuation method {val_method}.')
    return df


def dcf_valuation(df, n=10, rp=0.06, tg=0.06):
    beta = df['b_mkt']
    rf = df['RF']
    r = rf + beta * rp  # CAPM discount rate
    num_firms = df.shape[0]
    PV = np.zeros(num_firms)
    for i in range(n):
        ni_in_period_i = df[f'ni_{i+1}']
        PV += ni_in_period_i / (1 + r) ** i
        if i == n:
            PV += ni_in_period_i * (1 + tg) / (((1 + r) ** n) * (r - tg + 1e-10))
    return PV

rf_val = pd.read_csv('RF_ni_pred_for_val.csv')
rf_val['jdate'] = pd.to_datetime(rf_val['jdate'])

%%time
valdf = rf_val.merge(beta, on=['jdate', 'permno'])
valdf = perform_valuation(valdf, n=10)
valdf = valdf[['jdate', 'permno', 
               'year', 'PV']]
valdf.year = valdf.year + 1
val_crsp_mdf = valdf.merge(crsp, on=['year', 'permno'])
val_crsp_mdf['vt'] = val_crsp_mdf['PV']*1000

val_df1 = val_crsp_mdf[['vt', 'permno', 'ticker', 'me', 'year', 'jdate_crsp',
                        'logret_3', 'logret_6', 'logret_24', 'cumret_3', 'cumret_6', 'cumret_24',
                        'logret_12', 'logret_36', 'logret_60', 'cumret_12', 'cumret_36', 'cumret_60',
                         'termlogret', 'termcumret']]
val_df1.sort_values(['permno','jdate_crsp'], ascending = [True,False], inplace=True)

val_df2 = val_df1.copy()
for m in [3, 6, 12, 24, 36, 60]:
    val_df2[f'vt+{m}'] = val_df1.groupby('permno')['vt'].shift(m)
    val_df2[f'me+{m}'] = val_df1.groupby('permno')['me'].shift(m)

val_df3 = val_df2.copy()
for m in [3, 6, 12, 24, 36, 60]:
    val_df3['vt'] = np.log(val_df2.vt)
    val_df3[f'vt+{m}'] = np.log(val_df2[f'vt+{m}'].astype(float))
    val_df3['me'] = np.log(val_df2['me'].astype(float))
    val_df3[f'me+{m}'] = np.log(val_df2[f'me+{m}'].astype(float))
    val_df3[f'r_t+{m}'] = val_df3[f'logret_{m}'].fillna(val_df3['termlogret']+np.log(1-0.35))
    val_df3[f'r*_t+{m}'] = val_df3[f'vt+{m}'] -  val_df3[f'vt']
    val_df3[f'r_unchg+{m}'] = - (1-0.975) * (val_df3['me'] -  val_df3['vt'])
    val_df3[f'r_chg+{m}'] = 0.975 * ( (val_df3[f'me+{m}'] -  val_df3[f'vt+{m}']) - ((val_df3[f'me'] -  val_df3['vt'])) )
    

val_df3 = val_df3[1970<=val_df3.year]
val_df3 = val_df3.drop_duplicates(subset=['permno', 'jdate_crsp'])

val_df3['misp'] = val_df3['me'] - val_df3['vt']
val_df3 = val_df3.dropna(subset=['misp', 'r_t+60'])

val_df3[val_df3.ticker=='TLSA']

%%time
val_df3.to_csv('~/misp_data/decomp_dcf10y_RF_1970_2019.csv')
