==> rev_fcst/s8_fm_ret_on_gkx.ipynb <==
import numpy as np
import pandas as pd
import statsmodels as sms
import statsmodels.formula.api as smf
import matplotlib.pyplot as plt
from matplotlib.collections import LineCollection

import seaborn as sns # for data visualization
sns.set_style("whitegrid")

from dateutil.relativedelta import *
from pandas.tseries.offsets import *

pd.set_option('display.max_columns', None)

gkx = pd.read_csv('~/misp_data/gkx_mlfactors.csv')

gkx.tail()

gkx['date']=pd.to_datetime(gkx['DATE'], format='%Y%m%d')+MonthEnd(0)

ret_decomp_dcf5 = pd.read_csv('~/misp_data/decomp_dcf5y_RF_1970_2019.csv', index_col=0)
ret_decomp_dcf5['jdate_crsp'] = pd.to_datetime(ret_decomp_dcf5.jdate_crsp)

mdf5 = gkx.merge(ret_decomp_dcf5, left_on=['date', 'permno'], right_on=['jdate_crsp', 'permno'], how='inner')

vals = [
 'mvel1',
 'beta',
 'idiovol',
 'indmom',
 'mom1m',
 'mom6m',
 'mom12m',
 'mom36m',
 'pricedelay',
 'age',
 'agr',
 'bm',
 'bm_ia',
 'cashpr',
 'cfp',
 'cfp_ia',
 'chatoia',
 'chcsho',
 'depr',
 'divi',
 'egr',
 'ep',
 'grcapx',
 'hire',
 'invest',
 'lev',
 'mve_ia',
 'orgcap',
 'pchcurrat',
 'pchdepr',
 'pctacc',
 'quick',
 'rd',
 'rd_mve',
 'roic',
 'secured',
 'sp',
 'tang',
 'tb',
 'cash',
 'chtx',
 'cinvest',
 'nincr',
 'roaq',
 'roavol',
 'roeq',
 'rsup',
 'ms',
 'ill',
 'std_turn',]

# import mkl
# mkl.set_num_threads(5)

def ols_coef(x,formula):
    return smf.ols(formula,data=x).fit().params

def fama_macbeth_summary(p):
    s = p.describe().T
    s['std_error'] = s['std']/np.sqrt(s['count'])
    s['tstat'] = s['mean']/s['std_error']
    return s[['mean','std_error','tstat', 'count', 'std']]

def fama_macbeth_reg_panel(regdf, xname='misp', yname='ret', 
                     csname='permno', tsname='jdate_crsp'): 
    # Panel factor setting FM reg
    gamma_cs = (regdf.groupby(tsname).apply(ols_coef,f'{yname} ~ {xname}'))
    gamma_cs = gamma_cs.rename(columns={f"{xname}":f"{xname}"})
    gamma_cs=gamma_cs[[f"{xname}"]]
    return fama_macbeth_summary(gamma_cs)

mdf5['ret'] = np.exp(mdf5['r_t+60'])-1
mdf5['ret_ct'] = np.exp(mdf5['r*_t+60']-1)
mdf5['ret_chg'] = np.exp(mdf5['r_chg+60'])

%%time
res = []
for var in vals:
    regdf = mdf5[['jdate_crsp', 'permno', 'ticker', 'ret', var]].replace([np.inf, -np.inf], np.nan).dropna()
    print(f'Ret ~ {var}, 5Yr-DCF')
    results = fama_macbeth_reg_panel(regdf, xname=var, yname='ret')
#     print(results)
    res.append(results)
#     print('')

res_df  = pd.concat(res)

res_df.sort_values(['tstat']).tail()

%%time
res_ct = []
for var in vals:
    regdf = mdf5[['jdate_crsp', 'permno', 'ticker', 'ret_ct', var]].replace([np.inf, -np.inf], np.nan).dropna()
    print(f'{var}, retct')
    results = fama_macbeth_reg_panel(regdf, xname=var, yname='ret_ct')
#     print(results)
    res_ct.append(results)
#     print('')

res_ct_df  = pd.concat(res_ct)

res_ct_df.sort_values(['tstat']).tail()

%%time
res_chg = []
for var in vals:
    regdf = mdf5[['jdate_crsp', 'permno', 'ticker', 'ret_chg', var]].replace([np.inf, -np.inf], np.nan).dropna()
    print(f'{var}, retchg')
    results = fama_macbeth_reg_panel(regdf, xname=var, yname='ret_chg')
#     print(results)
    res_chg.append(results)
#     print('')

res_chg_df  = pd.concat(res_chg)

res_chg_df.sort_values(['tstat']).tail()

res_df = res_df.rename(columns={"mean":"mean_ret", "std_error":"std_error_ret",  "tstat":"tstat_ret"})
res_ct_df = res_ct_df.rename(columns={"mean":"mean_ct", "std_error":"std_error_ct",  "tstat":"tstat_ct"})
res_chg_df = res_chg_df.rename(columns={"mean":"mean_chg", "std_error":"std_error_chg",  "tstat":"tstat_chg"})

res_df = res_df[["mean_ret", "std_error_ret", "tstat_ret"]]
res_ct_df = res_ct_df[["mean_ct", "std_error_ct", "tstat_ct"]]
res_chg_df = res_chg_df[["mean_chg", "std_error_chg", "tstat_chg"]]

mm1 = res_df.merge(res_ct_df, left_index=True, right_index=True)
mm2 = mm1.merge(res_chg_df, left_index=True, right_index=True)

mm2.head()

mm2.to_csv('gkx_on_decomp_dcf5y_RF_1970_2019.csv')
